UPSTREAM: fuse: use mtime for readdir cache  verification
Store the modification time of the directory in the cache, obtained before
starting to fill the cache.

When reading the cache, verify that the directory hasn't changed, by
checking if current modification time is the same as the one stored in the
cache.

This only needs to be done when the current file position is at the
beginning of the directory, as mandated by POSIX.

Change-Id: I00fde78b4fbed1988bfde508cd5119d81a40fd86
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>