UPSTREAM: afs: Increase to 64-bit volume ID and  96-bit vnode ID for YFS
Increase the sizes of the volume ID to 64 bits and the vnode ID (inode
number equivalent) to 96 bits to allow the support of YFS.

This requires the iget comparator to check the vnode->fid rather than i_ino
and i_generation as i_ino is not sufficiently capacious.  It also requires
this data to be placed into the vnode cache key for fscache.

For the moment, just discard the top 32 bits of the vnode ID when returning
it though stat.

Change-Id: I9c7c9d261f9633d8e47688b49a75b6f9262b0cee
Signed-off-by: David Howells <dhowells@redhat.com>